requirements.txt
mysql-connector-python==9.3.0
mysqlclient==2.2.7
numpy==2.2.5
pandas==2.2.3
